Starting structuring crawler
